Changed byte type to int in a couple of classes in NQueen.GUI

+++ NQueen.GUI/ViewModels/MainViewModel.cs
@@ -45,7 +45,7 @@ public sealed partial class MainViewModel : ObservableObject, IDisposable
         _disposed = true;
     }
 
-    private void Initialize(byte boardSize = Utility.DefaultBoardSize,
+    private void Initialize(int boardSize = Utility.DefaultBoardSize,
         SolutionMode solutionMode = Utility.DefaultSolutionMode,
         DisplayMode displayMode = Utility.DefaultDisplayMode)
     {
@@ -76,7 +76,7 @@ public sealed partial class MainViewModel : ObservableObject, IDisposable
     {
         ObservableSolutions.Clear();
         Chessboard?.Squares.Clear();
-        BoardSize = byte.Parse(BoardSizeText);
+        BoardSize = int.Parse(BoardSizeText);
         NoOfSolutions = "0";
         ElapsedTimeInSec = $"{0,0:N1}";
         MemoryUsage = "0";
